src/backend/app/requirements.txt
fastapi
mysql-connector-python